backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyPDF2
python-docx
python-dotenv
sqlmodel
langchain_nvidia_ai_endpoints